A massive update:   Updated to a WinUsb version taken from:     https://github.com/old-dab/rtlsdr/
Hence no longer depends on libusb and Pthreads.
Other changes:
  *) Fixed the code-style using Astyle.
  *) Added more 'TRACE()' calls.
  *) etc, etc.

+++ externals/rtl-sdr/rtlsdr_i2c.h
@@ -3,7 +3,7 @@
 
 #define ARRAY_SIZE(arr) (sizeof(arr) / sizeof((arr)[0]))
 
-uint32_t rtlsdr_get_tuner_clock(void *dev);
+double rtlsdr_get_tuner_clock(void *dev);
 int rtlsdr_i2c_write_fn(void *dev, uint8_t addr, uint8_t reg, uint8_t *buf, int len);
 int rtlsdr_i2c_read_fn (void *dev, uint8_t addr, uint8_t reg, uint8_t *buf, int len);
